[Xamarin.Android.Build.Tasks] fix incremental builds with "stamp" files
Context: http://work.devdiv.io/707557

Many of our MSBuild targets declare `Outputs`, but don't always update
the timestamp of the output file. We commonly use `CopyIfChanged`
logic, so they won't write to the file if nothing changed.

This makes possible (actually quite easy) to get your incremental
build in a state where certain MSBuild targets *always* run.

~~ _BuildAdditionalResourcesCache ~~

This target usees the `$(_AndroidResourcePathsCache)` file as an
output, but the `<GetAdditionalResourcesFromAssemblies/>` MSBuild task
did *not* always update the timestamp. It uses
`XDocument.SaveIfChanged`.

I also removed an extraneous `<Message/>` here as you would already
see the information it's logging in a diagnostic build log. I also
moved the `Condition` for `$(DesignTimeBuild)` up to the target and
cleaned up the expression a bit. I had to update the
`AndroidUpdateResourcesTest.DesignTimeBuild` test for this change.

~~ _CleanIntermediateIfNuGetsChange ~~

This target was already correct, but I modified the location of the
stamp file to match our convention. I also removed the
`$(_AndroidNuGetStampFile)` property, which wasn't really needed.

I had to update the `BuildAfterUpgradingNuget` test here that was
asserting the stamp file exists, and it moved.

~~ _SetupDesignTimeBuildForCompile ~~

Needs to create `$(_AndroidStampDirectory)` if it does not exist, and
use a `Condition` with an `Exists` check. I found `<MakeDir/>` was
taking additional time (~10ms) without the `Exists` check? Perhaps it
is using a `try-catch`? We also don't need to check
`'$(DesignTimeBuild)' == 'true'` because this entire target doesn't
run in that case.

~~ _CreatePropertiesCache ~~

I removed a duplicate call to `<MakeDir/>` here, since a call to
create the `$(IntermediateOutputPath)\stamp` sub-directory will
automatically create `$(IntermediateOutputPath)` if it doesn't not
exist.

~~ _CreateNativeLibraryArchive ~~

This target used `__AndroidNativeLibraries__.zip` as an output, but
the logic in the `<CreateNativeLibraryArchive/>` target did *not*
always update the timestamp. `Files.ArchiveZipUpdate` uses
`CopyIfChanged`.  Added a `<Touch/>` so the zip file's timestamp is
always updated.

Other refactoring:
- Removed (weirdly named?) `_AddNativeLibraryArchiveToCompile` target
  and its usage of `CreateItem` which is deprecated.
- Added a simple `<ItemGroup/>` in `_CreateNativeLibraryArchive` to
  replace it
- Removed log messages from `CreateNativeLibraryArchive` and return
  `!Log.HasLoggedErrors`

~~ _CreateManagedLibraryResourceArchive ~~

This target used `__AndroidLibraryProjects__.zip` as an output, but
the logic in the `<CreateManagedLibraryResourceArchive/>` target did
*not* always update the timestamp. `Files.ArchiveZipUpdate` uses
`CopyIfChanged`. Added a `<Touch/>` so the zip file's timestamp is
always updated.

Other refactoring:
- Removed `_AddLibraryProjectsEmbeddedResourceToProject` target and
  its usage of `CreateItem` which is deprecated.
- Added a simple `<ItemGroup/>` in
  `_CreateManagedLibraryResourceArchive` to replace it
- The `CreateManagedLibraryResourceArchive` task had an unused
  `hasInvalidFilename` variable.
- `CreateManagedLibraryResourceArchive` should return
  `!Log.HasLoggedErrors` and removed extra log messages.
- Removed the `IsApplication` property, since the entire target is
  also checking this value.

~~ _CopyIntermediateAssemblies ~~

This target was already correct, but I modified the location of the
stamp file to match our convention.

~~ Other changes ~~

Expanded upon "MSBuild Best Practices" documentation. Fixed typos.