[build] <AcceptAndroidSdkLicense/> should not always run
On my machine, `<AcceptAndroidSdkLicense/>` is taking about ~1.5 sec
and runs even in a build with no changes.

We can run it as part of the `_UnzipFiles` target, removing the
`_AcceptAndroidSdkLicenses` target completely. We only need to
re-accept the licenses after new zip files are extracted.

This should save us about 1.5 seconds on any incremental build of
Xamarin.Android itself.