[Xamarin.Android.Build.Tasks] target to simplify @(FileWrites)
In our "MSBuild Best Practices" guide, the convention for using a
"stamp" file has a bit of boilerplate...

    <Target Name="_MyTarget"
        Inputs="..."
        Outputs="$(_AndroidStampDirectory)_MyTarget.stamp">
      <!--...-->
      <Touch Files="$(_AndroidStampDirectory)_MyTarget.stamp" AlwaysCreate="True" />
      <ItemGroup>
        <FileWrites Include="$(_AndroidStampDirectory)_MyTarget.stamp" />
      </ItemGroup>
    </Target>

It would be reasonably easy to mess this up.

If we add a new `_AddFilesToFileWrites` target that automatically adds
`$(_AndroidStampDirectory)*.stamp` to `FileWrites`, we can simplify
things:

    <Target Name="_MyTarget"
        Inputs="..."
        Outputs="$(_AndroidStampDirectory)_MyTarget.stamp">
      <!--...-->
      <Touch Files="$(_AndroidStampDirectory)_MyTarget.stamp" AlwaysCreate="True" />
    </Target>

This should be less prone to mistakes.