[Xamarin.Android.Build.Tasks] Add warning & error codes to GenerateJavaStubs
Context: https://github.com/xamarin/xamarin-android/issues/1560

Background discussion about the new wording for this particular warning
and error:

  * https://github.com/xamarin/xamarin-android/pull/2258#discussion_r222789036
  * https://github.com/xamarin/xamarin-android/pull/2258#discussion_r222798838

The new wording aims to be more actionable.  It is also more compact
than the old wording when there are multiple conflicts.

The `Dictionary` initializations follow the style used in [c362fe57][0].
Since these 2 new dictionaries should usually be empty, they are set to
an initial capacity of 0.  The current implementation of `Dictionary`
defaults to an initial capacity of 0, but the [documentation][1] doesn't
mention 0 specifically, so it seems reasonable to specify it.  Another
idea would be to initialize each dictionary only when a corresponding
conflict is found, but that might make the code trickier to read.

A possible future enhancement could be to move the warning into the
`ConvertCustomView` task so that it only appears when a managed type
name from an Android resource file matches more than one line in the
acw-map.  That way, if a user wants to solve the issue by using
`[Register]` attributes or assembly-qualified type names rather than
renaming the managed types, the build process will automatically stop
showing the warning.

[0]: https://github.com/xamarin/xamarin-android/commit/c362fe571dd96170a9e3e7e6134e72b6f957721d
[1]: https://docs.microsoft.com/dotnet/api/system.collections.generic.dictionary-2.-ctor