[build-tools] Output test results to console and txt on *nix.
We have a 2h timout on our *nix build bots.
Because the MSBuild tests can take *forever* to run,
they sometimes time out. This is mostly because
no output is written to the console.

This commit reworks the `RunNUnitTests` to use `tee` on
*nix to write the output from the test to a file. This is
useful because it also writes it to the console as well.

However, we have some tests will write Errors to the
output by design (they check for correct error messages).
So we need to use

	IgnoreStandardErrorWarningFormat="true"

to make sure that MSbuild does not think these are actual
errors. We will instead allow the NUnit ExitCode to make
that determination.