Move all MSBuild tests over to use `PackageReference`.
In a effort to reduce our dependency on external resources
we should not be relying on Nuget.exe to restore packages.
We should instead be using `msbuild /t:Restore.`
That only works with PackageReference though.

So lets move all of our Nuget references in the MSBuild
tests over to use that system.