[build-tools] Output test results to console & file on *nix (#2342)
Our Jenkins jobs use the [Build-timeout Plugin][0] so that they abort
if no output is written to the console within e.g. 2 hours.

Unfortunately, we still occasionally experience such timeout-caused
aborts (build failures), in part because the
`src/Xamarin.Android.Build.Tasks/Tests` unit tests take *forever* to
execute, during which nothing is written to the console.

This commit reworks the `RunNUnitTests` target to use **tee**(1) on
non-Windows platforms to write the output from the test to both a
file (as was done before) and to the console.  This helps ensure that
*something* is written to the console, thus "delaying" the abort.

However, we have some tests will write errors to the output by design
(they check for correct error messages):

	bash "/Users/builder/jenkins/workspace/xamarin-android-pr-builder/xamarin-android/build-tools/scripts/../../bin/TestDebug/parallel-targets.sh
	...
	error : unknown option '--no-crunch'.. This is the result of using `aapt` command line arguments with `aapt2`. The arguments are not compatible.

MSBuild "helpfully" converts this error message into an MSBuild error
which causes the overall build to fail:

	error MSB3073: The command "bash "…/xamarin-android/build-tools/scripts/../../bin/TestDebug/parallel-targets.sh"" exited with code -1.

(Side note: an exit code of a *negative* value -- `code -1` here --
means that "something" executed by the `<Exec/>` task wrote an error
message to stdout/stderr, and does *not* mean that the program
executed by the `<Exec/>` task exited with code `-1`, which isn't
even possible on Unix anyway...)

To prevent MSBuild from being "helpful" here, we need to set the
`Exec.IgnoreStandardErrorWarningFormat` to True:

	<Exec
	    Command="…"
	    IgnoreStandardErrorWarningFormat="True"
	/>

This will allow the *actual* exit code from the `Command` to
determine whether or not the `<Exec>` task succeeds or fails.

Additionally, the `parallel-targets.sh` invocation within the
`RunAllTests` target (91a39725) *also* needs to set
`Exec.IgnoreStandardErrorWarningFormat`=True, as `parallel-targets.sh`
*actually* executes the `RunNUnitTests` target, and thus is another
location that `error MSB3073` may be generated from (as seen above).

[0]: https://wiki.jenkins.io/display/JENKINS/Build-timeout+Plugin