[android-toolchain] split up targets + refactoring
Current when `android-toolchain.csproj` is built:
- A list of zip files are downloaded to `~/android-archives`.
- If *any* file is newer than a set of stamp files, all the
  destination directories are deleted.
- All the downloaded zip files are unzipped.

This is a bit inefficient, since updating something small like `ant`
or `depot_tools` will delete the Android SDK+NDK and rextract them.

I reworked the targets so that `_UnzipFiles` is split into:
- `_UnzipAndroidSdkItems`
- `_UnzipAndroidNdkItems`
- `_UnzipAntItems`
- `_UnzipChromeItems`

Other changes:
- Each new target uses a single stamp file. It doesn't make sense for
  these targets to "partially" build. If a single zip of the Android
  SDK changes, we need to unzip *everything*: not just the zip that is
  new.
- Removed use of `CreateItem`, a deprecated MSBuild task, in favor of
  `<ItemGroup/>`.
- Setup `%(HashHeader)` for each download group, even though it is
  only used by `depot_tools` right now.

As a result, the "stamp" files have moved. But after this PR,
`~/android-toolchain` will not get wiped nearly as often during
builds.