[Xamarin.Android.Build.Tasks] Fix Various `Lint` Issues
Fixes #2360, #2361, #2362, #2363

When running the `Lint` checks on an application we
provide a path to a `project` directory. This is
currently the `$(IntermediateOutputPath)`. The problem
with this is that it also includes the `lp` directory.
The `lp` directory contains ALL the resources and
Manifest files for the library projects. As a result
if you are referencing support packages ALL of those
items are checked too.

It turns out that most of the Manifest files shipped
with the libraries are not complete or will cause
`lint` errors or warnings.

So what we need to do is point the `lint` check at
the `$(IntermediateOutputPath)android`. This will restrict
the checks to the files which the Build system
produces (the java code, and the final app manifest).

This PR includes an updated unit test which tests all
the issues raised in the bug reports.