[Xamarin.Android.Build.Tasks] Fix Various `Lint` Issues (#2376)
Fixes: https://github.com/xamarin/xamarin-android/issues/2360
Fixes: https://github.com/xamarin/xamarin-android/issues/2361
Fixes: https://github.com/xamarin/xamarin-android/issues/2362
Fixes: https://github.com/xamarin/xamarin-android/issues/2363

When running the `<Lint/>` checks on an application we provide a path
to a `project` directory.  This was `$(IntermediateOutputPath)`.  The
problem with `$(IntermediateOutputPath)` is that it also includes the
`lp` directory, which contains ***all*** the resources and
`AndroidManifest.xml` files for all referenced library projects.  As
a result if you are referencing support packages all of those items
are checked too.

It turns out that most of the `AndroidManifest.xml` files shipped
with the support libraries are not complete or will cause `lint`
errors or warnings.

What we need to do is instead point the `lint` check at the
`$(IntermediateOutputPath)android` directory.  This will restrict the
checks to the files which the Build system produces (the java code,
and the final `AndroidManifest.xml` file).