[Xamarin.Android.Build.Tasks] Extract .so files to the `lp` directory (#2359)
To improve fast deployment even further we need to alter the way we
deal with native libraries in referenced assemblies.

Currently these libraries are extracted as part of the `<BuildApk/>`
task.  As a result if a native library is updated we need to build
the whole `.apk` again.

We should instead extract the `__AndroidNativeLibraries__.zip`
contents into the `$(IntermediateOutputDir)\lp\{assembly}\nl`
directory as part of the `<ResolveLibraryProjectImports/>` task.
This way the existing `<GetImportedLibraries/>` task will then pick
those `.so` files up and add them to the `@(NativeLibraries)` Item
Group.  This will then be picked up by `<BuildApk/>` task in a
release build.

However for a Commercial Debug build we can skip putting these files
in the `.apk` and Fast deploy them instead.