[plots] Do not mix Release and Release/Bundle measurements
Fixes https://github.com/xamarin/xamarin-android/issues/2400

With https://github.com/xamarin/xamarin-android/pull/1683 we started
to run and measure the apk tests with `BundleAssemblies=True` as
well. To not mix and repeast these measurements with other ones,
introduce `TestsFlavor` property in place of `TestsAotName` to
distinguish the apk tests runs when collecting and processing
measurements data.

It is similar to what we did with AOT before.

Example content of TestResult-Mono.Android_Tests-times.csv after the
fix:

    last-Release,JNI.init-Release,init-Release,NUnit.results-Release,last-Release-Aot,JNI.init-Release-Aot,init-Release-Aot,NUnit.results-Release-Aot,last-Release-Bundle,JNI.init-Release-Bundle,init-Release-Bundle,NUnit.results-Release-Bundle
    518,116,116,8347,647,86,86,7670,755,159,159,8416