[android-toolchain] mirror depot_tools in Azure storage (#2396)
In 4bb4b2e, we added a `%(HashHeader)` value for invalidating cached
downloads in `android-toolchain.csproj`.  Its primary use was for a
`depot_tools.zip` file, which is a dependency of building r8.

However...

Quite frequently we are seeing `depot_tools.zip` *change* and it gets
re-downloaded:

	Hash found using 'x-goog-hash', destination file changing to 'C:\…\android-archives\depot_tools-crc32c=JBYDJw==.zip'.
	Downloading `https://storage.googleapis.com/chrome-infra/depot_tools.zip` to `C:\…\android-archives\.depot_tools-crc32c=JBYDJw==.zip.download`.

Since we can't really count on a "version" of this file we are
downloading, I have mirrored it to the Azure storage account we are
already using for tests:

	https://xamjenkinsartifact.azureedge.net/mono-jenkins/xamarin-android/depot_tools/depot_tools-crc32c=JBYDJw==.zip

This URL will be reliable for our build, and we can update as needed.