[Xamarin.Android.Build.Tasks] Add Support to altering the AndroidManifest.xml
Context #1335

One of the things our users want to do is to alter the
merged `AndroidManifest.xml` before we package. This
PR adds support for a `AndroidManifestTransform` Build
action. This action allows users to supply an xslt
transform stylesheet to make changes to the manifest.

This transform will be applied after `GenerateJavaStubs`
has run.