[Xamarin.Android.Build.Tasks] Remove EmbeddedNativeLibrary from Mono.Data.Sqlite.
Fixes #2407

We currently embed the required native libraries for
Mono.Data.Sqlite and Mono.Posix into the assembly as
 an `EmbeddedResource` (__AndroidNativeLibraries__.zip).
This is then extracted as part of the build process
into the apk. This is the only actual change we have
from the vanilla versions of both of these assemblies.

This commit removes the `EmbeddedResource` items from
the assmeblies. It also updates sqlite-xamarin to place
its outputs into the

	`lib\xamarin-android\xbuild\Xamarin\Android\lib\{arch}\`

directories in the main output directory. This will make sure
we install them as part of the various installers we have.

We also update the MSBuild targets to check to see if these
assemblies are referenced and include the required `.so` files.