[Xamarin.Android.Build.Tasks] locate $(JdkJvmPath) only when needed
Fixes: https://github.com/xamarin/xamarin-android/issues/2409
Context: https://github.com/xamarin/xamarin-android/wiki/Build-Performance-Results

When doing the latest performance comparison, I noticed the
`<ResolveSdks/>` MSBuild task is a bit slower:

    * VS 15.9
     80 ms  ResolveSdks                                7 calls
    * VS 16.0 p2 (master)
    127 ms  ResolveSdks                                7 calls

For `$(AndroidGenerateJniMarshalMethods)` support, we needed to locate
a new `$(JdkJvmPath)` location.

Since `<ResolveSdks/>` finds `$(JdkJvmPath)` on first build, and
design-time builds, it is worth making sure this task doesn't get any
slower. The initial design-time build happens on "solution create",
which is one of the more important metrics the VS team measures.

So I split out the logic finding `$(JdkJvmPath)` and put it in a new
`<ResolveJdkJvmPath/>` MSBuild task.

Also setup a `Condition`, so it only runs if:

- Not a `$(DesignTimeBuild)`
- `$(AndroidGenerateJniMarshalMethods)` is enabled
- `$(JdkJvmPath)` is blank

This should save ~50ms on DTBs and first builds.