[Xamarin.Android.Build.Tasks] move $(AdbTarget) out of build.props
`$(IntermediateOutputPath)build.props`, or
`$(_AndroidBuildPropertiesCache)` is a file used as an `Input` to many
MSBuild targets. We write many *important* properties in this file as
a mechanism to trigger most of the build to rerun. So for example, if
you enabled proguard, the `$(AndroidEnableProguard)` MSBuild property
would re-run a lot of the build, such as `_CompileJava`,
`_CompileDex`, etc.

Right now there are two problematic properties in this file:
- `$(AdbTarget)`
- `$(AdbOptions)`

This means if you switched from a different Android deploy target:
emulator to device, or just a different device--it would run a full
build! `$(AdbTarget)` and `$(AdbOptions)` need to trigger MSBuild
targets related to "Fast Deployment" to re-run, but not everything.
What makes matters even worse, is sometimes design-time builds will
pass a blank `$(AdbTarget)`. This causes expensive targets to run
again.

I've moved these properties to a new file:
- `$(IntermediateOutputPath)adb.props` or
- `$(_AdbPropertiesCache)`

Then downstream in monodroid, there is a single line that needs to be
changed from `$(_AdbPropertiesCache)` to `$(_AdbPropertiesCache)`.
This will re-trigger the APK to be installed when devices change.