[Xamarin.Android.Build.Tasks] Fix Aar directory resolution. (#2424)
* [Xamarin.Android.Build.Tasks] Fix Aar directory resolution.

Fixed #2408

As part of the speed up process when extracting resources
from assemblies and Aar files, we dont re-extract if
the stamp file is newer than the assembly.

The problem with that was in the case of Aar files we
were NOT adding the `res` or `asset` directories to
the list of resolved directories when we skipped extraction.
As a result we got less `res` directories on a second
build.

This commit fixes that.

* Add Unit Test

* Fix Compile Error