Do not create override directory in Release builds
Fixes: https://github.com/xamarin/xamarin-android/issues/2053

Override directory is used only during development and, as such, is not created
by applications built in the `Release` configuration. The only exception to this
rule is when the release app is being profiled, in which case we do create the
override directory.

Recent migration to CMake for `src/monodroid` builds caused our release builds
of the XA runtime to not define the `RELEASE` macro and, in effect, the override
directory has always been created for release apps.

This commit rectifies it and makes the `src/monodroid` release
build **actually** work :)