[mono-sdks] <DownloadUri/> should not run during design-time builds (#2439)
Fixes: https://github.com/xamarin/xamarin-android/issues/2350

Most of the logic in `mono-runtimes.csproj` was happening during the
`CoreCompile` target.  The problem with this is that design-time
builds in Visual Studio runs `CoreCompile`.  The `Task.Wait()` in the
`<DownloadUri/>` task causes a hang within Visual Studio.

It turns out we could just completely remove the following target:

	<Target Name="CoreCompile"
	    DependsOnTargets="_Build">
	</Target>

`_Build` still runs appropriately. `_BuildUnlessCached` will call
`ForceBuild` or `_Build` depending if things are cached.