[mono-runtimes] Update @(MonoFacadeAssembly)
Fixes: https://github.com/xamarin/xamarin-android/issues/1879

Add the following Facade assemblies to `@(MonoFacadeAssembly)`:

  * `System.Buffers.dll`
  * `System.Memory.dll`
  * `System.Threading.Tasks.Extensions.dll`

Note: list was reordered to be sorted as per **sort**(1), and
generated with the following shell code on macOS:

	$ cd external/mono/mcs/class/Facades
	$ for d in `find . -depth 1 -type d | grep -v 'netstandard\|System.Drawing.Primitives\|System.Net.Http.Rtc' | sort -f` ; do
	  n=`basename "$d"`;
	  echo "    <MonoFacadeAssembly Include=\"$n.dll\" />";
	done | pbcopy

The `netstandard` and `System.Drawing.Primitives` directories are
excluded so that the comment and following `@(MonoFacadeAssembly)`
entries in `ProfileAssemblies.projitems` are still relevant.

The `System.Net.Http.Rtc` directory is excluded because it is only for
use on Desktop profiles.