[download-bundle] updated URL for Debug bundle
Context: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-debug/

To speed up build times, Xamarin.Android now builds Debug and Release
builds on different jobs on Jenkins.

The issue with this is that the Azure upload url will be of the form:

    https://xamjenkinsartifact.azureedge.net/mono-jenkins/${JOB_NAME}/bin/

So we need to check for `Debug` `$(Configuration)` and use
`xamarin-android-debug` in that case.